omega: System V leaf spill frames stay resident in the ABI red zone
A leaf function whose whole addressed frame extent is allocator spill
storage and fits the convention's red zone keeps every slot below the
unadjusted entry stack pointer: the layout records the resident extent,
the prologue commits nothing, and frame accesses resolve to signed
displacements. Residency is admitted only for System V AMD64 leaves with
no calls, no preservation storage, and no hosted or structural slots, so
no channel that cannot express a below-RSP coordinate ever observes one.
Independent replay reconstructs the residency, the committed-extent probe
roster, and the signed physical geometry; frame layout and encoding
identities bump so stale demand cannot alias the new realization.

+++ omega-rust/omega/backend/instruction_set_architectures/isa-aarch64/src/selected_form_encoding/memory/packed/tests.rs
@@ -68,7 +68,7 @@ fn packed_forms_access_exact_bytes_and_declare_all_scratch_writes() {
                         kind,
                         alternative,
                         &operands,
-                        offset,
+                        i64::from(offset),
                     )
                     .unwrap();
                     let effects = &encoded.footprint().encoded;
@@ -122,7 +122,7 @@ fn packed_forms_access_exact_bytes_and_declare_all_scratch_writes() {
                             kind,
                             alternative,
                             &operands,
-                            offset,
+                            i64::from(offset),
                             encoded.bytes()
                         )
                         .unwrap(),
@@ -187,7 +187,7 @@ fn packed_forms_reject_width_offset_alias_and_instruction_mutations() {
                     kind,
                     alternative,
                     &operands,
-                    offset
+                    i64::from(offset)
                 )
                 .is_err()
             );
